params: collapse ParamValue::Int into Float storage
ParamValue had separate `Float(f32)` and `Int(i32)` variants. The
asymmetry was the source of a class of silent bugs: any reader that
pattern-matched on `Float` only would fall through to the default
whenever a value was constructed as `Int` (e.g. from a JSON preset
declaring `{"type":"Int","value":N}`). The editor slider would move
and nothing visible would change — worst-class bug for live
performance.

The fix: numbers live in one cell. `ParamValue::Int` is removed,
all numerics use `Float(f32)`. `f32`'s 24-bit mantissa exactly
represents every integer in our param ranges, so storage is
lossless. `ParamType::Int` survives on `ParamDef` as a presentation
hint — the editor reads it to render whole-number steppers,
integer-formatted display, and round-on-drag — but storage doesn't
redundantly enforce it. The bug class is now structurally
impossible.

- `SerializedParamValue::Int { value }` stays in the wire schema as
  a read-only back-compat path. Old saved projects deserialise to
  `ParamValue::Float(value as f32)`; new saves emit `Float`.
- `ParamConvert::IntRound` still rounds to a whole number, now
  storing as `Float(rounded)`.
- New helpers `ParamValue::as_scalar()` / `as_u32_clamped(min)`
  centralise the coercion that ~30 primitive read sites duplicated.
- CI sweep test `int_typed_params_use_float_storage_and_coerce_cleanly`
  walks the primitive registry and asserts the contract per param.

Touches every primitive (write sites + read patterns), the graph
editor's drag-emit path, and the runtime conversion layer. Includes
the parallel mesh-typing pass adding `ItemKind` to `ArrayType` wires
(KnownItem trait + per-struct kinds) and renaming `LinePoint` →
`CurvePoint` for the centred-curve-space convention.

+++ crates/manifold-renderer/src/generators/json_graph_generator.rs
@@ -583,7 +583,7 @@ impl JsonGraphGenerator {
                 .unwrap_or(binding.default);
             let pv = match binding.convert {
                 ParamConvert::Float => ParamValue::Float(v),
-                ParamConvert::IntRound => ParamValue::Int(v.round() as i32),
+                ParamConvert::IntRound => ParamValue::Float(v.round()),
                 ParamConvert::BoolThreshold => ParamValue::Bool(v > 0.5),
                 ParamConvert::EnumRound => ParamValue::Enum(v.round().max(0.0) as u32),
             };
